Remove client input.
