--- packages/style/client.d.ts
@@ -1,2 +1,0 @@
-// Temporary!
-export * from './dts/index';
