--- packages/style/client.js
@@ -1,2 +1,0 @@
-// Temporary!
-module.exports = require('./lib/index.js');
